upgrade

+++ js/memo.js
@@ -8,6 +8,7 @@ window.addEventListener('DOMContentLoaded',function(){
         saveLocalStorage();
         delLocalStorage();
         selectTable();
+        allClearLocalStorage();
     }
 },false
 );
@@ -25,12 +26,16 @@ function saveLocalStorage(){
             window.alert("Key,Memo 入力必須です");
             return;
         }else{
+            let w_confirm= confirm(`LocalStorageに「${key}: ${value}」を保存しますか？`);
+            if(w_confirm===true){
+
             localStorage.setItem(key, value);
             viewStorage();
             let w_msg = "localStorageに" +key + " "+ value + "を保存しました";
             window.alert(w_msg);
             key="";
             value="";
+            }
         }
 
     },false
@@ -51,11 +56,15 @@ function viewStorage(){
         tr.appendChild(td1);
         tr.appendChild(td2);
         tr.appendChild(td3);
-
-        td1.innerHTML= "<input name='radio1' type= 'radio' >";
+        
+        td1.innerHTML= "<input name='chkbox1' type= 'checkbox' >";
         td2.innerHTML= w_key;
         td3.innerHTML= localStorage.getItem(w_key);
     }
+    $("#table1").tablesorter({
+        sortList: [[1, 0]] 
+    });
+    $("#table1").trigger("update");
 };
 
 function selectTable(){
@@ -63,47 +72,101 @@ function selectTable(){
     select.addEventListener("click",
         function(e) {
             e.preventDefault();
-            selectRadioBtn();
+            selectCheckBox("select");
         }, false
     )
 } 
 
-function selectRadioBtn(){
-    let w_sel= "0";
-    const radio1= document.getElementsByName('radio1');
-    const table1= document.getElementById("table1") 
-    for(let i=0; i< radio1.length; i++ ){
-        if(radio1[i].checked){
-            document.getElementById("textKey").value= table1.rows[i+1].cells[1].firstChild.data;
-            document.getElementById("textMemo").value= table1.rows[i+1].cells[2].firstChild.data;
-            return w_sel="1";
-        }
+function selectCheckBox(mode){
+    // let w_sel= "0";
+    let w_cnt =0;
+    const chkbox1= document.getElementsByName('chkbox1');
+    const table1 = document.getElementById("table1");
+    let w_textKey=""; 
+    let w_textMemo=""; 
+    for(let i=0; i< chkbox1.length; i++ ){
+        if(chkbox1[i].checked){
+            if(w_cnt===0){
+            w_textKey= table1.rows[i+1].cells[1].firstChild.data;
+            w_textMemo= table1.rows[i+1].cells[2].firstChild.data;
+            // return w_sel="1";
+            }
+        w_cnt++;
     }
-    window.alert("１つ選択して下さい");
-};
+    }
+    document.getElementById("textKey").value= w_textKey ;
+    document.getElementById("textMemo").value= w_textMemo ;
+
+    if(mode==="select"){
+        if(w_cnt===1){
+            return w_cnt;
+        }else{
+            window.alert("１つ選択して下さい");
+        }
+        }
+
+    if(mode==="del"){
+        if(w_cnt>=1){
+           return w_cnt;
+        }else{
+            window.alert("１つ選択して下さい");
+        }
+        }
+
+}
+
 
 function delLocalStorage() {
     const del = document.getElementById("del");
     del.addEventListener("click",
     function(e) {
         e.preventDefault();
-        let w_sel= "0";
-        w_sel= selectRadioBtn();
+        const chkbox1 =document.getElementsByName("chkbox1");
+        const table1 = document.getElementById("table1");
+        let w_cnt= 0;
+        w_cnt= selectCheckBox('del');
 
-        if(w_sel ==="1"){
-            const key= document.getElementById("textKey").value;
-            const value= document.getElementById("textMemo").value;
-            localStorage.removeItem(key);
+        const key= document.getElementById("textKey").value;
+        const value= document.getElementById("textMemo").value;
+    
+        if(w_cnt >=1){
+            let w_confirm = confirm(`LocalStorageから選択されている"${w_cnt}"を削除しますか？`);
+            if(w_confirm===true){
+            for(let i =0; i<chkbox1.length;i++){
+                if(chkbox1[i].checked){
+                    localStorage.removeItem(table1.rows[i+1].cells[1].firstChild.data);
+                }
+            }
+            // localStorage.removeItem(key);
             viewStorage();
-            let w_msg= "localStorageに" +key + " "+ value + "を削除しました";
+            let w_msg= `localStorageから${w_cnt} を削除しました`;
             window.alert(w_msg);
             document.getElementById("textKey").value = "";
             document.getElementById("textMemo").value= "";
         }
+    }
         
     },false
 
     );
+}
+
+function allClearLocalStorage(){
+    const allClear = document.getElementById("allClear");
+
+    allClear.addEventListener("click",
+    function(e){
+        e.preventDefault();
+        let w_confirm = confirm("LocalStorageのデータをすべて削除します。\nよろしですか。");
+        if(w_confirm ===true){
+            localStorage.clear();
+            viewStorage();
+            let w_msg= window.alert("LocalStorageのデータをすべて削除しました");
+            document.getElementById("textKey").value = "";
+            document.getElementById("textMemo").value= "";
+        }
+    }, false
+    )
 }
